[Refactor] move `menuRegistrant.ts` into `/browser/`
