+++ src/platform/menu/browser/menuRegistrant.ts
